tour_agency/requirements.txt
black==22.3.0
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
django-cors-headers==3.13.0
djangorestframework-gis==1.0
Pillow==9.1.1
olefile==0.46